[SYCL-MLIR][NFC] Move VersionConditionBuilder to TransformUtils for reuse
