+++ polygeist/lib/Dialect/Polygeist/Transforms/LICM.cpp
@@ -321,8 +321,6 @@ bool OperationSideEffects::conflictsWith(const Operation &other) const {
 /// This class represents an operation in a loop that is potentially invariant
 /// (if a condition is satisfied). The class tracks a list of operations (in the
 /// same loop) that need to be hoisted for this operation to be hoisted.
-using AccessorType = TypedValue<MemRefType>;
-using AccessorPairType = std::pair<AccessorType, AccessorType>;
 class LICMCandidate {
 public:
   LICMCandidate(Operation &op) : op(op) {}
@@ -350,163 +348,6 @@ private:
   /// Pairs of accessors that are required to not overlap for this operation to
   /// be invariant.
   SmallVector<AccessorPairType> requireNoOverlapAccessorPairs;
-};
-
-//===----------------------------------------------------------------------===//
-// VersionConditionBuilder
-//===----------------------------------------------------------------------===//
-
-class VersionConditionBuilder {
-public:
-  VersionConditionBuilder(
-      LoopLikeOpInterface loop,
-      ArrayRef<AccessorPairType> requireNoOverlapAccessorPairs)
-      : loop(loop), accessorPairs(requireNoOverlapAccessorPairs) {}
-
-  using SCFCondition = LoopVersionCondition::SCFCondition;
-  using AffineCondition = LoopVersionCondition::AffineCondition;
-
-  std::unique_ptr<LoopVersionCondition> createCondition() const {
-    OpBuilder builder(loop);
-    Location loc = loop.getLoc();
-    SCFCondition scfCond = createSCFCondition(builder, loc);
-    return std::make_unique<LoopVersionCondition>(scfCond);
-  }
-
-private:
-  /// Create a loop versioning condition suitable for scf::IfOp.
-  SCFCondition createSCFCondition(OpBuilder builder, Location loc) const {
-    auto GetMemref2PointerOp = [&](Value op) {
-      auto MT = cast<MemRefType>(op.getType());
-      return builder.create<polygeist::Memref2PointerOp>(
-          loc,
-          LLVM::LLVMPointerType::get(MT.getElementType(),
-                                     MT.getMemorySpaceAsInt()),
-          op);
-    };
-
-    Value condition;
-    for (const AccessorPairType &accessorPair : accessorPairs) {
-      Value begin1 = getSYCLAccessorBegin(accessorPair.first, builder, loc);
-      Value end1 = getSYCLAccessorEnd(accessorPair.first, builder, loc);
-      Value begin2 = getSYCLAccessorBegin(accessorPair.second, builder, loc);
-      Value end2 = getSYCLAccessorEnd(accessorPair.second, builder, loc);
-      auto beforeCond = builder.create<LLVM::ICmpOp>(
-          loc, LLVM::ICmpPredicate::ule, GetMemref2PointerOp(end1),
-          GetMemref2PointerOp(begin2));
-      auto afterCond = builder.create<LLVM::ICmpOp>(
-          loc, LLVM::ICmpPredicate::uge, GetMemref2PointerOp(begin1),
-          GetMemref2PointerOp(end2));
-      Value orOp = builder.create<arith::OrIOp>(loc, beforeCond, afterCond);
-      condition = condition
-                      ? builder.create<arith::AndIOp>(loc, condition, orOp)
-                      : orOp;
-    }
-    return condition;
-  }
-
-  template <typename OpTy>
-  static OpTy createMethodOp(OpBuilder builder, Location loc, Type resTy,
-                             ValueRange arguments, StringRef functionName,
-                             StringRef typeName) {
-    NamedAttrList attrs;
-    SmallVector<Type> argumentTypes;
-    for (Value argument : arguments)
-      argumentTypes.push_back(argument.getType());
-    attrs.set(mlir::sycl::SYCLDialect::getArgumentTypesAttrName(),
-              builder.getTypeArrayAttr(argumentTypes));
-    attrs.set(mlir::sycl::SYCLDialect::getFunctionNameAttrName(),
-              FlatSymbolRefAttr::get(builder.getStringAttr(functionName)));
-    attrs.set(mlir::sycl::SYCLDialect::getTypeNameAttrName(),
-              FlatSymbolRefAttr::get(builder.getStringAttr(typeName)));
-    return builder.create<OpTy>(loc, resTy, ValueRange(arguments), attrs);
-  }
-
-  static sycl::SYCLIDGetOp createSYCLIDGetOp(TypedValue<MemRefType> id,
-                                             unsigned index, OpBuilder builder,
-                                             Location loc) {
-    const Value indexOp = builder.create<arith::ConstantIntOp>(loc, index, 32);
-    const auto resTy = builder.getIndexType();
-    return createMethodOp<sycl::SYCLIDGetOp>(
-        builder, loc, MemRefType::get(ShapedType::kDynamic, resTy),
-        {id, indexOp}, "operator[]", "id");
-  }
-
-  static sycl::SYCLRangeGetOp createSYCLRangeGetOp(TypedValue<MemRefType> range,
-                                                   unsigned index,
-                                                   OpBuilder builder,
-                                                   Location loc) {
-    const Value indexOp = builder.create<arith::ConstantIntOp>(loc, index, 32);
-    const auto resTy = builder.getIndexType();
-    return createMethodOp<sycl::SYCLRangeGetOp>(
-        builder, loc, resTy, {range, indexOp}, "get", "range");
-  }
-
-  static sycl::SYCLAccessorGetRangeOp
-  createSYCLAccessorGetRangeOp(TypedValue<MemRefType> accessor,
-                               OpBuilder builder, Location loc) {
-    const auto accTy =
-        cast<sycl::AccessorType>(accessor.getType().getElementType());
-    const auto rangeTy = cast<sycl::RangeType>(
-        cast<sycl::AccessorImplDeviceType>(accTy.getBody()[0]).getBody()[1]);
-    return createMethodOp<sycl::SYCLAccessorGetRangeOp>(
-        builder, loc, rangeTy, accessor, "get_range", "accessor");
-  }
-
-  static sycl::SYCLAccessorSubscriptOp
-  createSYCLAccessorSubscriptOp(TypedValue<MemRefType> accessor,
-                                TypedValue<MemRefType> id, OpBuilder builder,
-                                Location loc) {
-    const auto accTy =
-        cast<sycl::AccessorType>(accessor.getType().getElementType());
-    const auto MT = cast<MemRefType>(
-        cast<LLVM::LLVMStructType>(accTy.getBody()[1]).getBody()[0]);
-    return createMethodOp<sycl::SYCLAccessorSubscriptOp>(
-        builder, loc, MT, {accessor, id}, "operator[]", "accessor");
-  }
-
-  static Value getSYCLAccessorBegin(TypedValue<MemRefType> accessor,
-                                    OpBuilder builder, Location loc) {
-    const auto accTy =
-        cast<sycl::AccessorType>(accessor.getType().getElementType());
-    const auto idTy = cast<sycl::IDType>(
-        cast<sycl::AccessorImplDeviceType>(accTy.getBody()[0]).getBody()[0]);
-    auto id = builder.create<memref::AllocaOp>(loc, MemRefType::get(1, idTy));
-    const Value zeroIndex = builder.create<arith::ConstantIndexOp>(loc, 0);
-    for (unsigned i = 0; i < accTy.getDimension(); ++i) {
-      Value idGetOp = createSYCLIDGetOp(id, i, builder, loc);
-      builder.create<memref::StoreOp>(loc, zeroIndex, idGetOp, zeroIndex);
-    }
-    return createSYCLAccessorSubscriptOp(accessor, id, builder, loc);
-  }
-
-  static Value getSYCLAccessorEnd(TypedValue<MemRefType> accessor,
-                                  OpBuilder builder, Location loc) {
-    const auto accTy =
-        cast<sycl::AccessorType>(accessor.getType().getElementType());
-    Value getRangeOp = createSYCLAccessorGetRangeOp(accessor, builder, loc);
-    auto range = builder.create<memref::AllocaOp>(
-        loc, MemRefType::get(1, getRangeOp.getType()));
-    const Value zeroIndex = builder.create<arith::ConstantIndexOp>(loc, 0);
-    builder.create<memref::StoreOp>(loc, getRangeOp, range, zeroIndex);
-    const auto idTy = cast<sycl::IDType>(
-        cast<sycl::AccessorImplDeviceType>(accTy.getBody()[0]).getBody()[0]);
-    auto id = builder.create<memref::AllocaOp>(loc, MemRefType::get(1, idTy));
-    const Value one = builder.create<arith::ConstantIndexOp>(loc, 1);
-    unsigned dim = accTy.getDimension();
-    for (unsigned i = 0; i < dim; ++i) {
-      Value idGetOp = createSYCLIDGetOp(id, i, builder, loc);
-      Value rangeGetOp = createSYCLRangeGetOp(range, i, builder, loc);
-      auto index = (i == dim - 1)
-                       ? rangeGetOp
-                       : builder.create<arith::SubIOp>(loc, rangeGetOp, one);
-      builder.create<memref::StoreOp>(loc, index, idGetOp, zeroIndex);
-    }
-    return createSYCLAccessorSubscriptOp(accessor, id, builder, loc);
-  }
-
-  mutable LoopLikeOpInterface loop;
-  ArrayRef<AccessorPairType> accessorPairs;
 };
 
 /// Return the accessor used by \p op if found, and nullptr otherwise.
