+++ polygeist/lib/Dialect/Polygeist/Utils/TransformUtils.cpp
@@ -8,7 +8,10 @@
 
 #include "mlir/Dialect/Polygeist/Utils/TransformUtils.h"
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
+#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
+#include "mlir/Dialect/Polygeist/IR/Ops.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
+#include "mlir/Dialect/SYCL/IR/SYCLOps.h"
 #include "llvm/ADT/TypeSwitch.h"
 
 using namespace mlir;
@@ -398,4 +401,141 @@ void LoopTools::guardLoop(LoopLikeOpInterface loop) const {
 void LoopTools::versionLoop(LoopLikeOpInterface loop,
                             const LoopVersionCondition &versionCond) const {
   LoopVersionBuilder(loop).versionLoop(versionCond);
+}
+
+//===----------------------------------------------------------------------===//
+// VersionConditionBuilder
+//===----------------------------------------------------------------------===//
+
+VersionConditionBuilder::SCFCondition
+VersionConditionBuilder::createSCFCondition(OpBuilder builder,
+                                            Location loc) const {
+  auto GetMemref2PointerOp = [&](Value op) {
+    auto MT = cast<MemRefType>(op.getType());
+    return builder.create<polygeist::Memref2PointerOp>(
+        loc,
+        LLVM::LLVMPointerType::get(MT.getElementType(),
+                                   MT.getMemorySpaceAsInt()),
+        op);
+  };
+
+  Value condition;
+  for (const AccessorPairType &accessorPair : accessorPairs) {
+    Value begin1 = getSYCLAccessorBegin(accessorPair.first, builder, loc);
+    Value end1 = getSYCLAccessorEnd(accessorPair.first, builder, loc);
+    Value begin2 = getSYCLAccessorBegin(accessorPair.second, builder, loc);
+    Value end2 = getSYCLAccessorEnd(accessorPair.second, builder, loc);
+    auto beforeCond = builder.create<LLVM::ICmpOp>(
+        loc, LLVM::ICmpPredicate::ule, GetMemref2PointerOp(end1),
+        GetMemref2PointerOp(begin2));
+    auto afterCond = builder.create<LLVM::ICmpOp>(loc, LLVM::ICmpPredicate::uge,
+                                                  GetMemref2PointerOp(begin1),
+                                                  GetMemref2PointerOp(end2));
+    Value orOp = builder.create<arith::OrIOp>(loc, beforeCond, afterCond);
+    condition =
+        condition ? builder.create<arith::AndIOp>(loc, condition, orOp) : orOp;
+  }
+  return condition;
+}
+
+template <typename OpTy>
+OpTy VersionConditionBuilder::createMethodOp(OpBuilder builder, Location loc,
+                                             Type resTy, ValueRange arguments,
+                                             StringRef functionName,
+                                             StringRef typeName) {
+  NamedAttrList attrs;
+  SmallVector<Type> argumentTypes;
+  for (Value argument : arguments)
+    argumentTypes.push_back(argument.getType());
+  attrs.set(mlir::sycl::SYCLDialect::getArgumentTypesAttrName(),
+            builder.getTypeArrayAttr(argumentTypes));
+  attrs.set(mlir::sycl::SYCLDialect::getFunctionNameAttrName(),
+            FlatSymbolRefAttr::get(builder.getStringAttr(functionName)));
+  attrs.set(mlir::sycl::SYCLDialect::getTypeNameAttrName(),
+            FlatSymbolRefAttr::get(builder.getStringAttr(typeName)));
+  return builder.create<OpTy>(loc, resTy, ValueRange(arguments), attrs);
+}
+
+sycl::SYCLIDGetOp
+VersionConditionBuilder::createSYCLIDGetOp(TypedValue<MemRefType> id,
+                                           unsigned index, OpBuilder builder,
+                                           Location loc) {
+  const Value indexOp = builder.create<arith::ConstantIntOp>(loc, index, 32);
+  const auto resTy = builder.getIndexType();
+  return createMethodOp<sycl::SYCLIDGetOp>(
+      builder, loc, MemRefType::get(ShapedType::kDynamic, resTy), {id, indexOp},
+      "operator[]", "id");
+}
+
+sycl::SYCLRangeGetOp
+VersionConditionBuilder::createSYCLRangeGetOp(TypedValue<MemRefType> range,
+                                              unsigned index, OpBuilder builder,
+                                              Location loc) {
+  const Value indexOp = builder.create<arith::ConstantIntOp>(loc, index, 32);
+  const auto resTy = builder.getIndexType();
+  return createMethodOp<sycl::SYCLRangeGetOp>(builder, loc, resTy,
+                                              {range, indexOp}, "get", "range");
+}
+
+sycl::SYCLAccessorGetRangeOp
+VersionConditionBuilder::createSYCLAccessorGetRangeOp(
+    TypedValue<MemRefType> accessor, OpBuilder builder, Location loc) {
+  const auto accTy =
+      cast<sycl::AccessorType>(accessor.getType().getElementType());
+  const auto rangeTy = cast<sycl::RangeType>(
+      cast<sycl::AccessorImplDeviceType>(accTy.getBody()[0]).getBody()[1]);
+  return createMethodOp<sycl::SYCLAccessorGetRangeOp>(
+      builder, loc, rangeTy, accessor, "get_range", "accessor");
+}
+
+sycl::SYCLAccessorSubscriptOp
+VersionConditionBuilder::createSYCLAccessorSubscriptOp(
+    TypedValue<MemRefType> accessor, TypedValue<MemRefType> id,
+    OpBuilder builder, Location loc) {
+  const auto accTy =
+      cast<sycl::AccessorType>(accessor.getType().getElementType());
+  const auto MT = cast<MemRefType>(
+      cast<LLVM::LLVMStructType>(accTy.getBody()[1]).getBody()[0]);
+  return createMethodOp<sycl::SYCLAccessorSubscriptOp>(
+      builder, loc, MT, {accessor, id}, "operator[]", "accessor");
+}
+
+Value VersionConditionBuilder::getSYCLAccessorBegin(
+    TypedValue<MemRefType> accessor, OpBuilder builder, Location loc) {
+  const auto accTy =
+      cast<sycl::AccessorType>(accessor.getType().getElementType());
+  const auto idTy = cast<sycl::IDType>(
+      cast<sycl::AccessorImplDeviceType>(accTy.getBody()[0]).getBody()[0]);
+  auto id = builder.create<memref::AllocaOp>(loc, MemRefType::get(1, idTy));
+  const Value zeroIndex = builder.create<arith::ConstantIndexOp>(loc, 0);
+  for (unsigned i = 0; i < accTy.getDimension(); ++i) {
+    Value idGetOp = createSYCLIDGetOp(id, i, builder, loc);
+    builder.create<memref::StoreOp>(loc, zeroIndex, idGetOp, zeroIndex);
+  }
+  return createSYCLAccessorSubscriptOp(accessor, id, builder, loc);
+}
+
+Value VersionConditionBuilder::getSYCLAccessorEnd(
+    TypedValue<MemRefType> accessor, OpBuilder builder, Location loc) {
+  const auto accTy =
+      cast<sycl::AccessorType>(accessor.getType().getElementType());
+  Value getRangeOp = createSYCLAccessorGetRangeOp(accessor, builder, loc);
+  auto range = builder.create<memref::AllocaOp>(
+      loc, MemRefType::get(1, getRangeOp.getType()));
+  const Value zeroIndex = builder.create<arith::ConstantIndexOp>(loc, 0);
+  builder.create<memref::StoreOp>(loc, getRangeOp, range, zeroIndex);
+  const auto idTy = cast<sycl::IDType>(
+      cast<sycl::AccessorImplDeviceType>(accTy.getBody()[0]).getBody()[0]);
+  auto id = builder.create<memref::AllocaOp>(loc, MemRefType::get(1, idTy));
+  const Value one = builder.create<arith::ConstantIndexOp>(loc, 1);
+  unsigned dim = accTy.getDimension();
+  for (unsigned i = 0; i < dim; ++i) {
+    Value idGetOp = createSYCLIDGetOp(id, i, builder, loc);
+    Value rangeGetOp = createSYCLRangeGetOp(range, i, builder, loc);
+    auto index = (i == dim - 1)
+                     ? rangeGetOp
+                     : builder.create<arith::SubIOp>(loc, rangeGetOp, one);
+    builder.create<memref::StoreOp>(loc, index, idGetOp, zeroIndex);
+  }
+  return createSYCLAccessorSubscriptOp(accessor, id, builder, loc);
 }
